+++ CHANGELOG.md
@@ -81,7 +81,7 @@ All notable, unreleased changes to this project will be documented in this file.
   - `pageType.availableAttributes`
 - Extend `AttributeEntityType` with `CATEGORY` and `COLLECTION`. You can now assign category and collection as a attribute reference.
 - Attribute values now expose the `referencedObject`, allowing for easier access to the linked entity.
-
+- You can now filter and search attribute choices using the new `where` and `search` fields on the `attribute.choices` query.
 
 ### Webhooks
 - Transaction webhooks responsible for processing payments can now return payment method details`, which will be associated with the corresponding transaction. See [docs](https://docs.saleor.io/developer/extending/webhooks/synchronous-events/transaction#response-4) to learn more.